Crate oak_restricted_kernel_sdk

Re-exports

pub use oak_attestation_verification;

Modules

attestation

Structs for providing attestation related logic such as getting an evidence.

channel

Provides functionality to communicate with host application over the communication channel.

crypto

Structs for signing and encryption using keys attested in the instance’s attestation evidence.

handler

testingtesting

Mock attestation evidence and crypto logic. Useful for testing where an attestation rooted in a real TEE may not be available.

utils

Various utilities like loggers, allocators, timers, etc.

Traits

Attester

Trait that provides the ability to build an attestation evidence.

EncryptionKeyHandle

Exposes the ability to derive a session key from the provided encapsulated private key, using a private key that has been endorsed in the Attestation Evidence.

Endorser

Trait that provides the ability to read attestation endorsements.

Signer

Attribute Macros

entrypoint

Marks a function as the entrypoint to an enclave app and sets up an conviences such an allocator, logger, panic handler.