Trait oak_restricted_kernel_sdk::attestation::EvidenceProvider

pub trait EvidenceProvider {
    // Required methods
    fn get_evidence(&self) -> &Evidence;
    fn get_encoded_event_log(&self) -> Option<&[u8]>;
}

Exposes the ability to read the Attestation Evidence. Note: Applications should only use the evidence to initially send it to the host application once, which then sends it to the clients. It is discouraged for enclave applications to operate directly with evidences.

Required Methods

fn get_evidence(&self) -> &Evidence

fn get_encoded_event_log(&self) -> Option<&[u8]>