Module oak_restricted_kernel_sdk::testing

Available on crate feature testing only.

Mock attestation evidence and crypto logic. Useful for testing where an attestation rooted in a real TEE may not be available.

Structs

MockAttester

Attester implementation that exposes mock evidence.

MockEncryptionKeyHandle

EncryptionKeyHandle implementation that using mock evidence and corresponding mock private keys.

MockSigner

Signer implementation that using mock evidence and corresponding mock private keys.