Trait oak_restricted_kernel_sdk::Attester

pub trait Attester: Send + Sync {
    // Required methods
    fn extend(&mut self, encoded_event: &[u8]) -> Result<(), Error>;
    fn quote(&self) -> Result<Evidence, Error>;
}

Trait that provides the ability to build an attestation evidence.

https://datatracker.ietf.org/doc/html/rfc9334#name-attester

Required Methods

fn extend(&mut self, encoded_event: &[u8]) -> Result<(), Error>

Add a new event to the evidence.

fn quote(&self) -> Result<Evidence, Error>

Generate a signed evidence containing all events previously provided with extend.

This function doesn’t take a mutable reference because quoting shouldn’t change the evidence. Evidence can only be updated via the extend function.