Crate oak_restricted_kernel_sdk

Modules

utils

Structs

FileDescriptorChannel

Channel that communicates over a file descriptor.

InstanceEncryptionKeyHandle

EncryptionKeyHandle implementation that using the instance’s evidence and corresponding private keys.

InstanceEvidenceProvider

EvidenceProvider implementation that exposes the instance’s evidence.

InstanceSigner

Signer implementation that using the instance’s evidence and corresponding private keys.

StderrLogger

log::Log implementation that outputs logs on standard error.

Traits

EncryptionKeyHandle

Exposes the ability to derive a session key from the provided encapsulated private key, using a private key that has been endorsed in the Attestation Evidence.

EvidenceProvider

Exposes the ability to read the Attestation Evidence. Note: Applications should only use the evidence to initially send it to the host application once, which then sends it to the clients. It is discouraged for enclave applications to operate directly with evidences.

Read

Simple no_std compatible equivalent of std::io::Read.

Signer

Exposes the ability to sign bytestrings using a private key that has been endorsed in the Attestation Evidence.

Write

Simple no_std compatible equivalent of std::io::Write.

Functions

alloc_error_handler

Provides a default implementation for alloc_error_handler attribute.

panic_handler

Provides a default implementation for panic_handler attribute.

start_blocking_server

Starts a blocking server that listens for requests on the provided channel and responds to them using the provided [micro_rpc::Transport].

Attribute Macros

entrypoint

Marks a function as the entrypoint to an enclave app and sets up an conviences such an allocator, logger, panic handler.