Attribute Macro oak_restricted_kernel_sdk::entrypoint

#[entrypoint]

Marks a function as the entrypoint to an enclave app and sets up an conviences such an allocator, logger, panic handler.

This macro assumes that crates using it have declared the no_std and no_main attributes, and the alloc_error_handler unstable feature of the rust compiler.

Examples

Filename: src/main.rs

#![no_std]
#![no_main]
#![feature(alloc_error_handler)]

extern crate alloc;

use oak_restricted_kernel_sdk::entrypoint;

#[entrypoint]
fn start_enclave_app() -> ! {
    // TODO(#0000): implement business logic below.
    unimplemented!();
}